+++ rts/motoko-rts/src/barriers.rs
@@ -1,5 +1,5 @@
-use motoko_rts_macros::*;
 use crate::{memory::Memory, types::Value};
+use motoko_rts_macros::*;
 
 #[incremental_gc]
 pub unsafe fn init_with_barrier<M: Memory>(_mem: &mut M, location: *mut Value, value: Value) {
