+++ rts/motoko-rts/src/gc.rs
@@ -12,7 +12,7 @@ use motoko_rts_macros::*;
 #[cfg(feature = "ic")]
 #[non_incremental_gc]
 unsafe fn should_do_gc(max_live: crate::types::Bytes<u64>) -> bool {
-    use crate::memory::ic::{HP, LAST_HP};
+    use crate::memory::ic::linear_memory::{HP, LAST_HP};
 
     // A factor of last heap size. We allow at most this much allocation before doing GC.
     const HEAP_GROWTH_FACTOR: f64 = 1.5;
