+++ rts/motoko-rts/src/gc/incremental.rs
@@ -69,20 +69,20 @@ static mut LAST_ALLOCATIONS: Bytes<u64> = Bytes(0u64);
 #[cfg(feature = "ic")]
 unsafe fn should_start() -> bool {
     use self::partitioned_heap::PARTITION_SIZE;
-    use crate::memory::ic;
+    use crate::memory::ic::partitioned_memory;
 
     const CRITICAL_HEAP_LIMIT: Bytes<u32> = Bytes(u32::MAX - 1024 * 1024 * 1024);
     const CRITICAL_GROWTH_THRESHOLD: f64 = 0.15;
     const NORMAL_GROWTH_THRESHOLD: f64 = 0.65;
 
-    let heap_size = ic::get_heap_size();
+    let heap_size = partitioned_memory::get_heap_size();
     let growth_threshold = if heap_size > CRITICAL_HEAP_LIMIT {
         CRITICAL_GROWTH_THRESHOLD
     } else {
         NORMAL_GROWTH_THRESHOLD
     };
 
-    let current_allocations = ic::get_total_allocations();
+    let current_allocations = partitioned_memory::get_total_allocations();
     debug_assert!(current_allocations >= LAST_ALLOCATIONS);
     let absolute_growth = current_allocations - LAST_ALLOCATIONS;
     let relative_growth = absolute_growth.0 as f64 / heap_size.as_usize() as f64;
@@ -91,18 +91,18 @@ unsafe fn should_start() -> bool {
 
 #[cfg(feature = "ic")]
 unsafe fn record_gc_start<M: Memory>() {
-    use crate::memory::ic;
-    LAST_ALLOCATIONS = ic::get_total_allocations();
+    use crate::memory::ic::partitioned_memory;
+    LAST_ALLOCATIONS = partitioned_memory::get_total_allocations();
 }
 
 #[cfg(feature = "ic")]
 unsafe fn record_gc_stop<M: Memory>() {
-    use crate::memory::ic;
+    use crate::memory::ic::{self, partitioned_memory};
 
-    let current_allocations = ic::get_total_allocations();
+    let current_allocations = partitioned_memory::get_total_allocations();
     debug_assert!(current_allocations >= LAST_ALLOCATIONS);
     let growth_during_gc = current_allocations - LAST_ALLOCATIONS;
-    let heap_size = ic::get_heap_size();
+    let heap_size = partitioned_memory::get_heap_size();
     let static_size = Bytes(ic::get_aligned_heap_base());
     debug_assert!(heap_size >= static_size);
     let dynamic_size = heap_size - static_size;
